ClientMsgs: Add reqSignatureLength to client request header
During implementation, I've found out there is not other easy way around
 - we must add this field.
If transaction signing is disabled - reqSignatureLength is 0.
On the next commit, we will add the needed changes to support this field
as part of messages ctors and handling on the replica side.
Also, to pass Apollo test I had to add some temporary changes to pass
tests while transaction signing is still disabled.

Lastly, bft_client is modified to fill this field after signing.